fix: Make discovery options is optional, refactor
