+++ src/discovery/mod.rs
@@ -1,4 +1,2 @@
 pub mod erc_4626;
 pub mod factory;
-
-pub mod discovery_options;
